product list successfully fetched and displyed for ranking, Category, and default

+++ app/src/main/java/com/pratham/headytest/ui/home/model/RankingUiModel.kt
@@ -3,5 +3,5 @@ package com.pratham.headytest.ui.home.model
 
 data class RankingUiModel(
     val name: String,
-    val id: Long?
+    val id: Int?
 )
